Create PlayerHospital, move AIHospital
