+++ CorsixTH/Lua/hospitals/ai_hospital.lua
@@ -1,0 +1,42 @@
+--[[ Copyright (c) 2020  Albert "Alberth" Hofkamp
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of
+this software and associated documentation files (the "Software"), to deal in
+the Software without restriction, including without limitation the rights to
+use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+of the Software, and to permit persons to whom the Software is furnished to do
+so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE. --]]
+
+class "AIHospital" (Hospital)
+
+---@type AIHospital
+local AIHospital = _G["AIHospital"]
+
+function AIHospital:AIHospital(competitor, ...)
+  self:Hospital(...)
+  if _S.competitor_names[competitor] then
+    self.name = _S.competitor_names[competitor]
+  else
+    self.name = "NONAME"
+  end
+  self.is_in_world = false
+end
+
+function AIHospital:spawnPatient()
+  -- TODO: Simulate patient
+end
+
+function AIHospital:logTransaction()
+  -- AI doesn't need a log of transactions, as it is only used for UI purposes
+end
