+++ CorsixTH/Lua/world.lua
@@ -32,6 +32,8 @@ corsixth.require("entities.humanoids.grim_reaper")
 corsixth.require("entities.humanoids.inspector")
 corsixth.require("staff_profile")
 corsixth.require("hospital")
+corsixth.require("hospitals.player_hospital")
+corsixth.require("hospitals.ai_hospital")
 corsixth.require("cheats")
 corsixth.require("epidemic")
 corsixth.require("calls_dispatcher")
